Rebase the branch to main (#75)
* Mofify slider (#68)

* Fix load codelist button

The load codelist button now works as intended. Importantly the search method changes also back to search if a users starts a new search after a codelist was loaded.

* Add launch app button to pkgdown site

This also removes the title bar of the app as it wont be needed anymore if we direct useres to the bennett github pkgdown site by default.

* Run `styler::style_pkg()`

* Change read.me wording: secondary care to inpatient secondary care (#74)

---------

Co-authored-by: Milan Wiedemann <milan.wiedemann@gmail.com>